backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
pandas>=2.0.0
langchain>=0.1.0
chromadb>=0.4.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.5.0